// src/lib/types.ts
export type ProductCategory = 
  | 'protein_powder'
  | 'mass_gainer'
  | 'pre_workout'
  | 'post_workout'
  | 'multivitamin'
  | 'omega_3'
  | 'creatine'
  | 'bcaa'
  | 'other';